+++ lognflow/__init__.py
@@ -2,12 +2,12 @@
 
 __author__ = 'Alireza Sadri'
 __email__ = 'arsadri@gmail.com'
-__version__ = '0.10.8'
+__version__ = '0.10.9'
 
 from .lognflow import lognflow
 from .logviewer import logviewer
 from .printprogress import printprogress
-from .plt_utils import plt_colorbar, plot_gaussian_gradient
+from .plt_utils import plt_colorbar, plot_gaussian_gradient, plt_imshow
 from .utils import (
     select_directory, select_file, repr_raw, replace_all, 
     text_to_object, stack_to_frame, stacks_to_frames, ssh_system)